Part one completed
The for loop iterates through the numbers from 1 to 100.
Inside the loop, I checked the following conditions=>
If the number is divisible by both 3 and 5 using the modulo operator %,

If the number is divisible by 3, we log "Fizz".
If the number is divisible by 5, we log "Buzz".
If the number is not divisible by either 3 or 5, we simply log the number.